RIA-7213_internal_ada_appeal_decided_dismissed_document: InternalAdaDecisionsReasonsLetterAllowedGenerator(and Dismissed) refactored into one generic class with conditions, tests fixed as well
