+++ src/functionalTest/resources/scenarios/RIA-7212-internal-ada-send-decision-and-reasons-allowed-document.json
@@ -1,6 +1,5 @@
 {
   "description": "RIA-7212 Internal ADA send decision and reasons - allowed - document PDF",
-  "disabled": "true",
   "request": {
     "uri": "/asylum/ccdAboutToSubmit",
     "credentials": "Judge",
